:recycle: refactor: CrudRepository 제너레이터 <VO 타입, Primary Key 타입>

+++ backend/src/main/java/com/clothes/perst/persistance/MusinsaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MusinsaRepository extends JpaRepository<MusinsaVO, Long> {
+public interface MusinsaRepository extends JpaRepository<MusinsaVO, Integer> {
 
     // 필수 : Gender, Style
     // 선택 : musinsa_personal, color
